minor refactor and generated user service tests

+++ TodoApp/src/main/java/com/sgz/TodoApp/services/UserService.java
@@ -1,10 +1,7 @@
 package com.sgz.TodoApp.services;
 
-import com.google.common.collect.Sets;
-import com.sgz.TodoApp.entities.Role;
 import com.sgz.TodoApp.entities.User;
 import com.sgz.TodoApp.exceptions.*;
-import com.sgz.TodoApp.repos.RoleRepo;
 import com.sgz.TodoApp.repos.UserRepo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.password.PasswordEncoder;
@@ -18,21 +15,18 @@ import java.util.Optional;
 public class UserService {
 
     private final UserRepo userRepo;
-    private final RoleRepo roleRepo;
     private final PasswordEncoder passwordEncoder;
 
     @Autowired
-    public UserService(UserRepo userRepo, RoleRepo roleRepo, PasswordEncoder passwordEncoder) {
+    public UserService(UserRepo userRepo, PasswordEncoder passwordEncoder) {
         this.userRepo = userRepo;
-        this.roleRepo = roleRepo;
         this.passwordEncoder = passwordEncoder;
     }
 
-    public User createUser(User toAdd) throws InvalidEntityException, InvalidNameException, InvalidAuthorityException {
+    public User createUser(User toAdd) throws InvalidEntityException, InvalidNameException {
         validate(toAdd);
         checkExistsByUsername(toAdd.getUsername());
 
-        toAdd.setRoles(Sets.newHashSet(getRoleByAuthority("USER")));
         toAdd.setPassword(passwordEncoder.encode(toAdd.getPassword()));
 
         return userRepo.save(toAdd);
@@ -40,6 +34,7 @@ public class UserService {
 
     public List<User> getAllUsers() throws NoItemsException {
         List<User> allUsers = userRepo.findAll();
+
         if (allUsers.isEmpty()) {
             throw new NoItemsException("No Items");
         }
@@ -60,13 +55,11 @@ public class UserService {
         return toGet.get();
     }
 
-    public User editUser(User toEdit, int authId) throws InvalidEntityException, InvalidIdException, InvalidAuthorityException, InvalidNameException, AccessDeniedException {
-        isAuthoritzed(toEdit.getId(), authId);
+    public User editUser(User toEdit, int authId) throws InvalidEntityException, InvalidIdException, AccessDeniedException {
+        checkAuthorization(toEdit.getId(), authId);
         validate(toEdit);
         checkExists(toEdit.getId());
-        checkExistsByUsername(toEdit.getUsername());
 
-        toEdit.setRoles(Sets.newHashSet(getRoleByAuthority("USER")));
         toEdit.setPassword(passwordEncoder.encode(toEdit.getPassword()));
 
         return userRepo.save(toEdit);
@@ -74,6 +67,7 @@ public class UserService {
 
     public User getUserById(int id) throws InvalidIdException {
         Optional<User> toGet = userRepo.findById(id);
+
         if (!toGet.isPresent()) {
             throw new InvalidIdException("Invalid Id");
         }
@@ -81,38 +75,21 @@ public class UserService {
         return toGet.get();
     }
 
-    public void deleteUserById(int writeId, int authId) throws InvalidIdException, InvalidNameException, InvalidEntityException, AccessDeniedException {
-        isAuthoritzed(writeId, authId);
+    public void deleteUserById(int writeId, int authId) throws InvalidIdException, AccessDeniedException {
+        checkAuthorization(writeId, authId);
         checkExists(writeId);
         userRepo.deleteById(writeId);
     }
 
-    private Role getRoleByAuthority(String authority) throws InvalidAuthorityException, InvalidEntityException {
-        if (authority == null || authority.trim().isEmpty()) {
-            throw new InvalidEntityException("Name is invalid");
-        }
-
-        Optional<Role> toGet = roleRepo.findByAuthority(authority);
-        if (!toGet.isPresent()) {
-            throw new InvalidAuthorityException("Authority not found");
-        }
-
-        return toGet.get();
-    }
-
     private void checkExistsByUsername(String username) throws InvalidNameException {
-        if (userRepo.existsByUsername(username)) {
-            throw new InvalidNameException("Name already exists");
-        }
+        if (userRepo.existsByUsername(username)) throw new InvalidNameException("Name already exists");
     }
 
-    private void checkExists(int id) throws InvalidIdException, InvalidNameException, InvalidEntityException {
-        if (!userRepo.existsById(id)) {
-            throw new InvalidIdException("Invalid Id");
-        }
+    private void checkExists(int id) throws InvalidIdException {
+        if (!userRepo.existsById(id)) throw new InvalidIdException("Invalid Id");
     }
 
-    private void isAuthoritzed(int writeId, int authId) throws AccessDeniedException {
+    private void checkAuthorization(int writeId, int authId) throws AccessDeniedException {
         if(writeId != authId) throw new AccessDeniedException("Access Denied");
     }
 
@@ -123,6 +100,8 @@ public class UserService {
                 || toUpsert.getPassword().trim().isEmpty()
                 || toUpsert.getPassword().trim().length() > 20
                 || !toUpsert.getPassword().matches("^(?=.*[0-9])(?=.*[a-z])(?=.*[A-Z])(?=.*[@#$%^&+=])(?=\\S+$).{8,}$")
+                || toUpsert.getRoles() == null
+                || toUpsert.getRoles().isEmpty()
         ) {
             throw new InvalidEntityException("Invalid Entity");
         }
